move db into supabase and add minor changes for other files , added env.example too
