+++ backend/.env.example
@@ -1,0 +1,6 @@
+# Mock values for backend environment variables
+CLERK_PUBLISHABLE_KEY=pk_test_example_clerk_publishable_key
+SUPABASE_URL=https://example.supabase.co
+SUPABASE_ANON_KEY=example_supabase_anon_key
+LOCAL_HOST_URL=http://localhost
+PORT=3001
