+++ backend/src/server.ts
@@ -1,8 +1,7 @@
 import express from 'express';
 import cors from 'cors';
 import dotenv from 'dotenv';
-
-import supabase from './db';
+import supabase from '../supabase/db';
 import UserRouter from './routes/userRoutes';
 // import other routers like TripRouter, NotificationRouter if needed
 
@@ -20,17 +19,24 @@ app.use('/user', UserRouter);
 
 // ✅ Start server
 const startServer = async () => {
-  // If you want to test connection, you can add a Supabase health check here (optional)
-  console.log('Checking Supabase connectivity...');
-  // Test query to check connection
-  const { error } = await supabase.from('users').select('*').limit(1);
-  if (error) {
-    console.error('❌ Supabase connection failed:', error.message);
-    return;
+  try {
+    // If you want to test connection, you can add a Supabase health check here (optional)
+    console.log('Checking Supabase connectivity...');
+    // Test query to check connection
+    const { error } = await supabase.from('users').select('*').limit(1);
+    if (error) {
+      console.error('❌ Supabase connection failed:', error.message);
+      return;
+    }
+    app.listen(PORT, () => {
+      console.log(`✅ Server running at ${LOCAL_HOST_URL}:${PORT}`);
+    });
+  } catch (err : any) {
+    // Catches both Supabase errors and unexpected runtime errors
+    console.error('🚨 Failed to start server:', err.message || err);
+    // Exit with failure code so process managers (PM2, Docker, etc.) know something went wrong
+    process.exit(1);
   }
-  app.listen(PORT, () => {
-    console.log(`✅ Server running at ${LOCAL_HOST_URL}${PORT}`);
-  });
 };
 
 startServer();
