+++ backend/tsconfig.json
@@ -2,13 +2,13 @@
   "compilerOptions": {
     "target": "ES2020", // Modern JS target
     "module": "commonjs", // Node.js compatible modules
-    "rootDir": "src", // Source files
+    "rootDir": ".", // Source files
     "outDir": "dist", // Compiled JS output
     "strict": true, // Enable all strict type-checking
     "esModuleInterop": true, // Allow `import x from 'y'` with commonjs
     "skipLibCheck": true, // Skip type checking for node_modules
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["**/*.ts", "**/*.tsx"],
+  "include": ["**/*.ts", "**/*.tsx", "supabase/**/*.ts"],
   "exclude": ["node_modules", "dist"]
 }
